+++ src/exactly_lib/help/entities/actors/objects/file_interpreter_actor.py
@@ -1,32 +1,26 @@
 from typing import List
 
 from exactly_lib.actors import file_interpreter as actor
-from exactly_lib.common.help.syntax_contents_structure import SyntaxElementDescription, InvokationVariant
+from exactly_lib.common.help.syntax_contents_structure import SyntaxElementDescription, InvokationVariant, \
+    invokation_variant_from_args
 from exactly_lib.definitions import instruction_arguments, formatting
 from exactly_lib.definitions.cross_ref.app_cross_ref import SeeAlsoTarget
 from exactly_lib.definitions.entity import concepts, syntax_elements
 from exactly_lib.definitions.entity.actors import FILE_INTERPRETER_ACTOR
-from exactly_lib.definitions.test_case import actor as help_texts
 from exactly_lib.help.entities.actors.contents_structure import ActorDocumentation
-from exactly_lib.help.entities.actors.objects.common import ARGUMENT_SYNTAX_ELEMENT, \
+from exactly_lib.help.entities.actors.objects.common import \
     SINGLE_LINE_PROGRAM_ACT_PHASE_CONTENTS_SYNTAX_INITIAL_PARAGRAPH, ActPhaseDocumentationSyntaxBase
 from exactly_lib.help.program_modes.common.render_syntax_contents import invokation_variants_content
 from exactly_lib.help.render import doc_utils
 from exactly_lib.section_document.syntax import LINE_COMMENT_MARKER
 from exactly_lib.test_case_utils.documentation.relative_path_options_documentation import path_element_2
 from exactly_lib.util.cli_syntax.elements import argument as a
-from exactly_lib.util.cli_syntax.render import cli_program_syntax
 from exactly_lib.util.textformat.structure import document as doc
-from exactly_lib.util.textformat.structure.core import ParagraphItem
 from exactly_lib.util.textformat.structure.structures import section_contents
 from exactly_lib.util.textformat.textformat_parser import TextParser
 
 
 class FileInterpreterActorDocumentation(ActorDocumentation):
-    CL_SYNTAX_RENDERER = cli_program_syntax.CommandLineSyntaxRenderer()
-
-    ARG_SYNTAX_RENDERER = cli_program_syntax.ArgumentInArgumentDescriptionRenderer()
-
     def __init__(self):
         super().__init__(FILE_INTERPRETER_ACTOR)
         self._tp = TextParser({
@@ -53,6 +47,7 @@ class FileInterpreterActorDocumentation(ActorDocumentation):
         return [
             syntax_elements.ACT_INTERPRETER_SYNTAX_ELEMENT.cross_reference_target,
             syntax_elements.PATH_SYNTAX_ELEMENT.cross_reference_target,
+            syntax_elements.PROGRAM_ARGUMENT_SYNTAX_ELEMENT.cross_reference_target,
         ]
 
 
@@ -62,26 +57,21 @@ DOCUMENTATION = FileInterpreterActorDocumentation()
 class ActPhaseDocumentationSyntax(ActPhaseDocumentationSyntaxBase):
     def __init__(self):
         self.file = instruction_arguments.FILE_ARGUMENT
-        self.argument = a.Named(help_texts.ARGUMENT)
-        fm = {
-            'shell_syntax_concept': formatting.concept_(concepts.SHELL_SYNTAX_CONCEPT_INFO),
-        }
-        super().__init__(TextParser(fm))
+        super().__init__(TextParser(dict()))
 
     def invokation_variants(self) -> List[InvokationVariant]:
         executable_arg = a.Single(a.Multiplicity.MANDATORY, self.file)
-        optional_arguments_arg = a.Single(a.Multiplicity.ZERO_OR_MORE, self.argument)
         return [
-            InvokationVariant(self._cl_syntax_for_args([executable_arg,
-                                                        optional_arguments_arg])),
+            invokation_variant_from_args([
+                executable_arg,
+                syntax_elements.PROGRAM_ARGUMENT_SYNTAX_ELEMENT.zero_or_more,
+            ]),
         ]
 
     def syntax_element_descriptions(self) -> List[SyntaxElementDescription]:
         return [
             path_element_2(actor.RELATIVITY_CONFIGURATION,
                            self._parser.fnap(_SOURCE_FILE_SYNTAX_ELEMENT)),
-            SyntaxElementDescription(self.argument.name,
-                                     self._parser.fnap(ARGUMENT_SYNTAX_ELEMENT)),
         ]