+++ CHANGELOG.md
@@ -20,6 +20,7 @@ with exception of a "0." prefix until version 1 is released.
  - Syntax of type expressions - Precedences of `||` and `&&`
  - Type `files-matcher` - rename `-prune` -> `-with-pruned`
  - Interpreter actors - interpreter `EXECUTABLE [ARGUMENT]...` requires `EXECUTABLE` to be a file
+ - Actor - `file interpreter` - arguments to source file are `PROGRAM-ARGUMENT`
  - Exit codes - rename `IMPLEMENTATION_ERROR` -> `INTERNAL_ERROR`
  - Python compatibility: Increased to Python v 3.5.9
  - Instruction `conf`/`actor` - Remove "-" prefix from actor names